Move httpapi to server/httpapi
